thay doi lai cau truc hoan thien cac bussiness logic co ban 4 entity Category, Order, orrderItem, Product

+++ new_toy_store/src/main/java/com/example/new_toy_store/order/application/dto/response/OrderResponse.java
@@ -5,19 +5,19 @@ import java.util.List;
 
 public class OrderResponse {
 
-    private Long id;
+    private Integer id;
     private String status;
     private LocalDateTime createdAt;
     private List<OrderItemResponse> items;
 
-    public OrderResponse(Long id, String status, LocalDateTime createdAt, List<OrderItemResponse> items) {
+    public OrderResponse(Integer id, String status, LocalDateTime createdAt, List<OrderItemResponse> items) {
         this.id = id;
         this.status = status;
         this.createdAt = createdAt;
         this.items = items;
     }
 
-    public Long getId() { return id; }
+    public Integer getId() { return id; }
 
     public String getStatus() { return status; }
